requirements.txt
asgiref==3.4.1
pytz==2021.1
sqlparse==0.4.2

Django==3.1.6
Pillow==8.3.2
psycopg2-binary==2.9.1

django-import-export==2.6.1
dj-database-url==0.5.0
gunicorn==20.1.0
whitenoise==5.3.0

